Add Portainer Agent, add node placement constraints
I haven't often worked with multi-node clusters, and have gotten used to
bind mounting files for simpler management.  Introducing a secondary
node to this deployment to help deal with load has challenged that
habit.  For now, I have added constraints to run many of the services on
the manager node, though that isn't exactly scalable.